requirements.txt
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-openai==1.1.10
langchain-text-splitters==1.1.1
chromadb==1.5.2
pypdf==6.7.5
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
beautifulsoup4==4.12.3
